Add ScalaDoc, refactor model package structure, refactor GameState to MatchState, itemsInRooms into geography
