--- core/src/main/scala/io/github/scalaquest/core/model/GameState.scala
@@ -1,8 +1,0 @@
-package io.github.scalaquest.core.model
-
-trait GameState[I <: Model#Item] {
-  def player: Player[I]
-  def ended: Boolean
-  def rooms: Set[Room]
-  def itemsInRooms: Map[Room, Set[I]]
-}
