+++ core/src/main/scala/io/github/scalaquest/core/model/Message.scala
@@ -1,3 +1,6 @@
 package io.github.scalaquest.core.model
 
+/**
+ * A representation a single line of output to render to the user at the end of the pipeline round.
+ */
 trait Message
