+++ core/src/main/scala/io/github/scalaquest/core/model/Room.scala
@@ -1,18 +1,27 @@
 package io.github.scalaquest.core.model
 
-import io.github.scalaquest.core.model.Direction.Direction
-import io.github.scalaquest.core.model.std.StdRoom
-
-object Direction extends Enumeration {
-  type Direction = Value
-  val NORTH, SOUTH, WEST, EAST, UP, DOWN = Value
-}
+import io.github.scalaquest.core.model.Room.Direction
 
 trait Room {
   def name: String
   def neighbors(direction: Direction): Option[Room]
 }
 
+case class StdRoom(name: String, _neighbors: () => Map[Direction, Room]) extends Room {
+  override def neighbors(direction: Direction): Option[Room] = _neighbors() get direction
+}
+
 object Room {
   def apply(name: String, neighbors: () => Map[Direction, Room]): Room = StdRoom(name, neighbors)
+
+  sealed trait Direction
+
+  object Direction {
+    case object North extends Direction
+    case object South extends Direction
+    case object East  extends Direction
+    case object West  extends Direction
+    case object Up    extends Direction
+    case object Down  extends Direction
+  }
 }
