+++ core/src/main/scala/io/github/scalaquest/core/model/behaviorBased/impl/package.scala
@@ -1,0 +1,7 @@
+package io.github.scalaquest.core.model.behaviorBased
+
+/**
+ * The package makes available to the storyteller a ready-to-use implementation of
+ * [[BehaviorBasedModel]], exploits (mixing in) all the constructs from the [[common]] package.
+ */
+package object impl {}
