+++ core/src/main/scala/io/github/scalaquest/core/model/behaviorBased/package.scala
@@ -1,0 +1,9 @@
+package io.github.scalaquest.core.model
+
+/**
+ * The package makes available to the storyteller an implementation of [[Model]] following a
+ * <b>behavior-based fashion</b>. This is a style that enables the "decoration" of [[Model.Item]] s
+ * and the [[Model.Ground]] by the use of the <b>Behavior</b>: a sort of composable plugin that
+ * matches dynamically various action-item-state scenarios.
+ */
+package object behaviorBased {}
