+++ core/src/main/scala/io/github/scalaquest/core/model/package.scala
@@ -1,0 +1,6 @@
+package io.github.scalaquest.core
+
+/**
+ * The model contains all concepts modifiable by the storyteller, to build his/her story.
+ */
+package object model {}
