--- core/src/main/scala/io/github/scalaquest/core/model/std/StdRoom.scala
@@ -1,8 +1,0 @@
-package io.github.scalaquest.core.model.std
-
-import io.github.scalaquest.core.model.Direction.Direction
-import io.github.scalaquest.core.model.Room
-
-case class StdRoom(name: String, _neighbors: () => Map[Direction, Room]) extends Room {
-  override def neighbors(direction: Direction): Option[Room] = _neighbors() get direction
-}
